Move identity verification and renewal forms to components, prevent accessing renewal form without identity verification
